refactoring events to its files
